+++ ui/src/commonMain/kotlin/quantityChanger/QunatityChanger.kt
@@ -9,7 +9,7 @@ import androidx.compose.animation.slideOutHorizontally
 import androidx.compose.animation.togetherWith
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
-import androidx.compose.foundation.clickable
+import androidx.compose.foundation.gestures.detectTapGestures
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.padding
@@ -20,6 +20,7 @@ import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -27,9 +28,12 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Outline
 import androidx.compose.ui.graphics.Shape
+import androidx.compose.ui.input.pointer.pointerInput
 import androidx.compose.ui.unit.Density
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.LayoutDirection
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.launch
 
 @Composable
 fun QuantityChanger(
@@ -41,7 +45,7 @@ fun QuantityChanger(
 ) {
     var previousQuantity by remember { mutableStateOf(currentQuantity) }
     val transitionSpecPair = remember {
-        val animationDuration = 300
+        val animationDuration = 250
         val tweenSpecInt = tween<IntOffset>(durationMillis = animationDuration)
         val tweenSpecFloat = tween<Float>(durationMillis = animationDuration)
 
@@ -60,6 +64,7 @@ fun QuantityChanger(
             transitionSpecPair.second
         }
     }
+    val scope = rememberCoroutineScope()
     Row(
         modifier = modifier,
         verticalAlignment = Alignment.CenterVertically,
@@ -79,7 +84,20 @@ fun QuantityChanger(
                 ).padding(
                     quantityChangerConfig.actionPaddingValues
                 ).clip(previousIconShape)
-                .clickable { onQuantityChange(currentQuantity - 1) },
+                .pointerInput(Unit) {
+                    detectTapGestures(
+                        onPress = {
+                            val job = scope.launch {
+                                while (true) {
+                                    onQuantityChange(previousQuantity - 1)
+                                    delay(250)
+                                }
+                            }
+                            tryAwaitRelease()
+                            job.cancel()
+                        }
+                    )
+                },
             imageVector = quantityChangerConfig.prevIcon,
             contentDescription = null,
             tint = quantityChangerConfig.actionIconTint,
@@ -118,7 +136,20 @@ fun QuantityChanger(
                 .padding(
                     quantityChangerConfig.actionPaddingValues
                 ).clip(nextIconShape)
-                .clickable { onQuantityChange(currentQuantity + 1) },
+                .pointerInput(Unit) {
+                    detectTapGestures(
+                        onPress = {
+                            val job = scope.launch {
+                                while (true) {
+                                    onQuantityChange(previousQuantity + 1)
+                                    delay(250)
+                                }
+                            }
+                            tryAwaitRelease()
+                            job.cancel()
+                        }
+                    )
+                },
             imageVector = quantityChangerConfig.nextIcon,
             contentDescription = null,
             tint = quantityChangerConfig.actionIconTint,